Init implementation of entities, not functional yet
