+++ srcopsmetrics/entities/ContentFile.py
@@ -1,0 +1,93 @@
+#!/usr/bin/env python3
+# Copyright (C) 2020 Dominik Tuchyna
+#
+# This file is part of SrcOpsMetrics.
+#
+# SrcOpsMetrics is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# SrcOpsMetrics is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with SrcOpsMetrics.  If not, see <http://www.gnu.org/licenses/>.
+
+"""ContentFile entity class."""
+
+import logging
+
+from github.ContentFile import GithubContentFile
+from github.PaginatedList import PaginatedList
+from voluptuous.schema_builder import Schema
+
+from srcopsmetrics.entities.BaseEntity import BaseEntity
+from srcopsmetrics.enums import EntityTypeEnum
+from srcopsmetrics.iterator import KnowledgeAnalysis
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class ContentFile(BaseEntity):
+    """GitHub ContentFile entity."""
+
+    entity_name = "ContentFile"
+
+    entity_schema = Schema({"name": str, "path": str, "content": str})  # TODO: Adjust content type
+
+    entities_schema = Schema({str: entity_schema})
+
+    def __init__(self, repository):
+        """Initialize with repo and prev knowledge."""
+        self.stored = {}
+        self.repository = repository
+        self.prev_knowledge = None
+
+    def analyse(self) -> PaginatedList:
+        """Override :func:`~BaseEntity.analyse`."""
+        _LOGGER.info("-------------Content Files Analysis-------------")
+
+        # TODO: Extend to all types of files. Currently only README are considered.
+        # TODO: Add all types of README extensions available
+        content_file_text = ""
+        for file_name in ["README.md", "README.rst"]:
+
+            try:
+                content_file = self.repository.get_contents(file_name)
+                file_path = content_file.path
+                encoded = content_file.decoded_content
+                # TODO: Adjust because most of the files are not text
+                content_file_text = encoded.decode("utf-8")
+            except Exception as e:
+                _LOGGER.info("%r not found for: %r" % (file_name, self.repository.full_name))
+                _LOGGER.warning(e)
+
+            if content_file_text:
+                with KnowledgeAnalysis(
+                    entity_type=EntityTypeEnum.CONTENT_FILE.value,
+                    new_entities=[[file_name, file_path, content_file_text]],
+                    accumulator=self.prev_knowledge,
+                    store_method=self.store,
+                ) as analysis:
+                    accumulated = analysis.store()
+                break
+
+        if not content_file_text:
+            return None
+
+        return accumulated
+
+    def store(self, file_content: GithubContentFile):
+        """Override :func:`~BaseEntity.store`."""
+        self.stored["content_files"] = {
+            "name": file_content[0],
+            "path": file_content[1],
+            "content": file_content[2],
+        }
+
+    def stored_entities(self):
+        """Override :func:`~BaseEntity.stored_entities`."""
+        return self.stored
