+++ srcopsmetrics/github_knowledge.py
@@ -20,9 +20,9 @@
 import logging
 import os
 from pathlib import Path
-from typing import Any, Dict, Generator, List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
-from github import ContentFile, Github, Issue, PaginatedList, PullRequest
+from github import ContentFile, Github, PaginatedList
 from github.Repository import Repository
 
 from srcopsmetrics.enums import EntityTypeEnum
@@ -32,8 +32,6 @@ from srcopsmetrics.storage import KnowledgeStorage
 _LOGGER = logging.getLogger(__name__)
 
 _GITHUB_ACCESS_TOKEN = os.getenv("GITHUB_ACCESS_TOKEN")
-
-ISSUE_KEYWORDS = {"close", "closes", "closed", "fix", "fixes", "fixed", "resolve", "resolves", "resolved"}
 
 STANDALONE_LABELS = {"size"}
 
@@ -118,58 +116,6 @@ class GitHubKnowledge:
         return [label for label in labels if label not in STANDALONE_LABELS]
 
     @staticmethod
-    def search_for_references(body: str) -> Generator[str, None, None]:
-        """Return generator for iterating through referenced IDs in a comment."""
-        if body is None:
-            return
-
-        message = body.split(" ")
-        for idx, word in enumerate(message):
-            if word.replace(":", "").lower() not in ISSUE_KEYWORDS:
-                return
-
-            _LOGGER.info("      ...found keyword referencing issue")
-            referenced_issue_number = message[idx + 1]
-            if referenced_issue_number.startswith("https"):
-                # last element of url is always the issue number
-                ref_issue = referenced_issue_number.split("/")[-1]
-            elif referenced_issue_number.startswith("#"):
-                ref_issue = referenced_issue_number.replace("#", "")
-            else:
-                _LOGGER.info("      ...referenced issue number absent")
-                _LOGGER.debug("      keyword message: %s" % body)
-                return
-
-            if not referenced_issue_number.isnumeric():
-                _LOGGER.info("      ...referenced issue number in incorrect format")
-                return
-
-            _LOGGER.info("      ...referenced issue number: %s" % ref_issue)
-            yield ref_issue
-
-    def get_referenced_issues(self, pull_request: PullRequest) -> List[str]:
-        """Scan all of the Pull Request comments and get referenced issues.
-
-        Arguments:
-            pull_request {PullRequest} -- Pull request for which the referenced
-                                        issues are extracted
-
-        Returns:
-            List[str] -- IDs of referenced issues within the Pull Request.
-
-        """
-        issues_referenced = []
-        for comment in pull_request.get_issue_comments():
-            for id in self.search_for_references(comment.body):
-                issues_referenced.append(id)
-
-        for id in self.search_for_references(pull_request.body):
-            issues_referenced.append(id)
-
-        _LOGGER.debug("      referenced issues: %s" % issues_referenced)
-        return issues_referenced
-
-    @staticmethod
     def get_only_new_entities(old_data: Dict[str, Any], new_data: PaginatedList) -> PaginatedList:
         """Get new entities (whether PRs or other Issues).
 
@@ -206,155 +152,6 @@ class GitHubKnowledge:
             # we count by the num of words in comment
             interactions[comment.user.login] += len(comment.body.split(" "))
         return interactions
-
-    def store_issue(self, issue: Issue, data: Dict[str, Dict[str, Any]]):
-        """Extract required information from issue and store it to the current data.
-
-        This is targeted only for issues that are not Pull Requests.
-
-        Arguments:
-            issue {Issue} -- Issue (that is not PR).
-            data {Dict[str, Union[str, int]])} -- Dictionary where the issue will be stored.
-
-        """
-        if issue.pull_request is not None:
-            return  # we analyze issues and prs differentely
-
-        labels = [label.name for label in issue.get_labels()]
-
-        data[str(issue.number)] = {
-            "created_by": issue.user.login,
-            "created_at": int(issue.created_at.timestamp()),
-            "closed_by": issue.closed_by.login if issue.closed_by is not None else None,
-            "closed_at": int(issue.closed_at.timestamp()) if issue.closed_at is not None else None,
-            "labels": self.get_non_standalone_labels(labels),
-            "interactions": self.get_interactions(issue.get_comments()),
-        }
-
-        # TODO: think about saving comments
-        # would it be valuable?
-
-    def analyse_issues(
-        self, repository: Repository, prev_knowledge: Dict[str, Any], is_local: bool = False
-    ) -> Optional[Dict[str, Any]]:
-        """Analyse of every closed issue in repository.
-
-        Arguments:
-            repository {Repository} -- currently the PyGithub lib is used because of its functionality
-                                    ogr unfortunatelly did not provide enough to properly analyze issues
-
-            prev_knowledge {Dict[str, Any]} -- previous knowledge stored
-
-        """
-        _LOGGER.info("-------------Issues (that are not PR) Analysis-------------")
-
-        current_issues = [issue for issue in repository.get_issues(state="all") if issue.pull_request is None]
-        new_issues = self.get_only_new_entities(prev_knowledge, current_issues)
-
-        if len(new_issues) == 0:
-            return None
-
-        with KnowledgeAnalysis(
-            entity_type=EntityTypeEnum.ISSUE.value,
-            new_entities=new_issues,
-            accumulator=prev_knowledge,
-            store_method=self.store_issue,
-        ) as analysis:
-            accumulated = analysis.store()
-
-        return accumulated
-
-    @staticmethod
-    def extract_pull_request_review_requests(pull_request: PullRequest) -> List[str]:
-        """Extract features from requested reviews of the PR.
-
-        GitHub understands review requests rather as requested reviewers than actual
-        requests.
-
-        Arguments:
-            pull_request {PullRequest} -- PR of which we can extract review requests.
-
-        Returns:
-            List[str] -- list of logins of the requested reviewers
-
-        """
-        requested_users = pull_request.get_review_requests()[0]
-
-        extracted = []
-        for user in requested_users:
-            extracted.append(user.login)
-        return extracted
-
-    @staticmethod
-    def extract_pull_request_reviews(pull_request: PullRequest) -> Dict[str, Dict[str, Any]]:
-        """Extract required features for each review from PR.
-
-        Arguments:
-            pull_request {PullRequest} -- Pull Request from which the reviews will be extracted
-
-        Returns:
-            Dict[str, Dict[str, Any]] -- dictionary of extracted reviews. Each review is stored
-
-        """
-        reviews = pull_request.get_reviews()
-        _LOGGER.info("  -num of reviews found: %d" % reviews.totalCount)
-
-        results = {}
-        for idx, review in enumerate(reviews, 1):
-            _LOGGER.info("      -analysing review no. %d/%d" % (idx, reviews.totalCount))
-            results[str(review.id)] = {
-                "author": review.user.login,
-                "words_count": len(review.body.split(" ")),
-                "submitted_at": int(review.submitted_at.timestamp()),
-                "state": review.state,
-            }
-        return results
-
-    def store_pull_request(self, pull_request: PullRequest, results: Dict[str, Dict[str, Any]]):
-        """Analyse pull request and save its desired features to results.
-
-        Arguments:
-            pull_request {PullRequest} -- PR that is going to be inspected and stored.
-            results {Dict[str, Dict[str, Any]]} -- dictionary where all the currently
-                                                PRs are stored and where the given PR
-                                                will be stored.
-
-        """
-        commits = pull_request.commits
-        # TODO: Use commits to extract information.
-        # commits = [commit for commit in pull_request.get_commits()]
-
-        created_at = int(pull_request.created_at.timestamp())
-        closed_at = int(pull_request.closed_at.timestamp()) if pull_request.closed_at is not None else None
-        merged_at = int(pull_request.merged_at.timestamp()) if pull_request.merged_at is not None else None
-
-        closed_by = pull_request.as_issue().closed_by.login if pull_request.as_issue().closed_by is not None else None
-
-        labels = [label.name for label in pull_request.get_labels()]
-
-        # Evaluate size of PR
-        pull_request_size = None
-        if labels:
-            pull_request_size = self.get_labeled_size(labels)
-
-        if not pull_request_size:
-            lines_changes = pull_request.additions + pull_request.deletions
-            pull_request_size = self.assign_pull_request_size(lines_changes=lines_changes)
-
-        results[str(pull_request.number)] = {
-            "size": pull_request_size,
-            "labels": self.get_non_standalone_labels(labels),
-            "created_by": pull_request.user.login,
-            "created_at": created_at,
-            "closed_at": closed_at,
-            "closed_by": closed_by,
-            "merged_at": merged_at,
-            "commits_number": commits,
-            "referenced_issues": self.get_referenced_issues(pull_request),
-            "interactions": self.get_interactions(pull_request.get_issue_comments()),
-            "reviews": self.extract_pull_request_reviews(pull_request),
-            "requested_reviewers": self.extract_pull_request_review_requests(pull_request),
-        }
 
     def store_content_file(self, file_content: ContentFile, results: Dict[str, Dict[str, Any]]):
         """Analyse pull request and save its desired features to results.
@@ -415,36 +212,6 @@ class GitHubKnowledge:
 
         if not content_file_text:
             return None
-
-        return accumulated
-
-    def analyse_pull_requests(
-        self, repository: Repository, prev_knowledge: Dict[str, Any], is_local: bool = False
-    ) -> Optional[Dict[str, Any]]:
-        """Analyse every closed pull_request in repository.
-
-        Arguments:
-            repository {Repository} -- currently the PyGithub lib is used because of its functionality
-                                    ogr unfortunatelly did not provide enough to properly analyze issues
-
-            prev_knowledge {Dict[str, Any]} -- previous knowledge stored
-
-        """
-        _LOGGER.info("-------------Pull Requests Analysis (including its Reviews)-------------")
-
-        current_pulls = repository.get_pulls(state="all")
-        new_pulls = self.get_only_new_entities(prev_knowledge, current_pulls)
-
-        if len(new_pulls) == 0:
-            return None
-
-        with KnowledgeAnalysis(
-            entity_type=EntityTypeEnum.PULL_REQUEST.value,
-            new_entities=new_pulls,
-            accumulator=prev_knowledge,
-            store_method=self.store_pull_request,
-        ) as analysis:
-            accumulated = analysis.store()
 
         return accumulated
